# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FieldForge API
  version: 0.3.0
  description: |
    Multitenant ERP API for field services (US).
    Protected routes require `Authorization: Bearer <token>` and `X-Tenant-ID` header.

    ## Cross-cutting middleware

    - **Rate limiting:** Public routes (`/auth/signup`, `/auth/login`) are limited per client IP.
      Authenticated routes are limited per tenant. Responses include `X-RateLimit-Limit`,
      `X-RateLimit-Remaining`, and `X-RateLimit-Reset`. HTTP 429 includes `Retry-After`.
    - **Idempotency:** `POST` requests on protected routes accept optional `Idempotency-Key`.
      Replays within 24h return the cached response with `Idempotency-Replayed: true`.
    - **Domain events:** Mutations publish to the transactional outbox (`domain_events`);
      a background worker marks events `processed` after dispatch.
servers:
  - url: https://api.fieldforge.com/api/v1
    description: Production
  - url: http://localhost:8081/api/v1
    description: Local API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer }
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer, description: Unix timestamp when the window resets }
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, example: HTTP_429 }
                  message: { type: string, example: rate limit exceeded }
...
paths:
...
  /auth/signup:
    post:
      summary: Create tenant and owner account
      tags: [Auth]
      description: Rate limited per client IP (default 20 req/min).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_name, email, password, industry_pack]
              properties:
                company_name: { type: string }
                email: { type: string, format: email }
                password: { type: string }
                industry_pack: { type: string }
                plan_id: { type: string }
      responses:
        "201":
          description: Account created with JWT
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/login:
    post:
      summary: Authenticate user
      tags: [Auth]
      description: Rate limited per client IP (default 20 req/min).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: JWT token
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/me:
    get:
      summary: Current user profile
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/TenantID" }]
      responses:
        "200":
          description: User profile
...
  /nav:
    get:
      summary: Aggregated navigation items from enabled plugins
      tags: [System]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/TenantID" }]
      responses:
        "200":
          description: Nav items
  /billing/checkout:
    post:
      summary: Create Stripe checkout session (mock when debug.mock_stripe=true)
      tags: [Billing]
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/TenantID" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                plan_id: { type: string }
                success_url: { type: string }
                cancel_url: { type: string }
      responses:
        "200":
          description: Checkout session with checkout_url
  /billing/webhook:
    post:
      summary: Stripe webhook receiver
      tags: [Billing]
      parameters:
        - name: Stripe-Signature
          in: header
          schema: { type: string }
      responses:
        "200":
          description: Event received
...
  /public/contact:
    post:
      summary: Submit marketing contact form (stub)
      tags: [Platform]
      description: >
        Accepts demo and sales inquiries from the marketing site. In development,
        submissions are logged when debug.features.skip_email_send is true.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email, company, inquiry, message]
              properties:
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string, format: email }
                company: { type: string }
                inquiry:
                  type: string
                  enum: [demo, sales, support, enterprise]
                message: { type: string }
      responses:
        "201":
          description: Submission accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  reference_id: { type: string, format: uuid }
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /public/marketing-content:
    get:
      summary: Blog listings and industry case studies for marketing site
      tags: [Platform]
      description: Returns content from config/marketing-content.yaml (stub CMS).
      responses:
        "200":
          description: Marketing content payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  blog:
                    type: object
                    properties:
                      posts:
                        type: array
                        items:
                          type: object
                  case_studies:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: object
        "429":
          $ref: "#/components/responses/RateLimited"
...